move bsff sign modals with other signatures
