+++ front/src/Apps/Dashboard/Validation/bsff/queries.ts
